+++ src/Damage/IvanoviSALEDamageModel.hh
@@ -83,9 +83,13 @@ public:
   virtual void enforceBoundaries(State<Dimension>& state,
                                  StateDerivatives<Dimension>& derivs) override;
 
-  // An optional hook to initialize once when the problem is starting up.
-  virtual void initializeProblemStartup(DataBase<Dimension>& dataBase) override;
-
+  // A second optional method to be called on startup, after Physics::initializeProblemStartup has
+  // been called.
+  // One use for this hook is to fill in dependendent state using the State object, such as
+  // temperature or pressure.
+  virtual void initializeProblemStartupDependencies(DataBase<Dimension>& dataBase,
+                                                    State<Dimension>& state,
+                                                    StateDerivatives<Dimension>& derivs) override;
   //............................................................................
   // Accessors for state
   double minPlasticFailure() const;