+++ src/GSPH/GenericRiemannHydro.hh
@@ -77,9 +77,13 @@ public:
                           const StateDerivatives<Dimension>& derivs,
                           const Scalar currentTime) const override;
 
-  // Tasks we do once on problem startup.
-  virtual
-  void initializeProblemStartup(DataBase<Dimension>& dataBase) override;
+  // A second optional method to be called on startup, after Physics::initializeProblemStartup has
+  // been called.
+  // One use for this hook is to fill in dependendent state using the State object, such as
+  // temperature or pressure.
+  virtual void initializeProblemStartupDependencies(DataBase<Dimension>& dataBase,
+                                                    State<Dimension>& state,
+                                                    StateDerivatives<Dimension>& derivs) override;
 
   // Register the state Hydro expects to use and evolve.
   virtual