+++ RELEASE_NOTES.md
@@ -17,6 +17,7 @@ Notable changes include:
     * Fixed bug with ConstantBoundary in the presence of porosity with the new porosity models introduced in v2024.01.00.
     * Updating header lists for including Spheral modules in external projects.
     * Adding effective viscous pressure back to FSISPH.
+    * Initial volumes for damage models were incorrectly not taking into account pore space when computing failure statistics for seeding flaws.  Fixed.
 
 Version v2024.01.00 -- Release date 2024-01-19
 ==============================================